Simple Readme Update
Since this is just a readme update, it will be automatically committed to main. After this point any and all updates will occur in the usual format of creating a new branch, putting in a pull request, and committing format. There will most likely also be a development branch as well to adhere to git flow, even though this is an entirely personal repo.